+++ lua/autorun/ttt_deadringer_sound.lua
@@ -1,0 +1,19 @@
+AddCSLuaFile()
+
+sound.Add({
+    name = 'deadringer_activate',
+    channel = CHAN_WEAPON,
+    volume = 1.0,
+    level = 75,
+    pitch = 100,
+    sound = 'buttons/blip1.wav'
+})
+
+sound.Add({
+    name = 'deadringer_deactivate',
+    channel = CHAN_WEAPON,
+    volume = 1.0,
+    level = 75,
+    pitch = 73,
+    sound = 'buttons/blip1.wav'
+})